Address Evan's comments
